requirements.txt
scrapling[fetchers]
requests
pillow
customtkinter
lxml
fastapi
uvicorn
pydantic
psycopg2-binary
